# openapi/analytics-admin.openapi.json
{
  "components": {
...
    "securitySchemes": {
      "metrics": { "bearerFormat": "JWT", "scheme": "bearer", "type": "http" },
      "session": {
        "in": "cookie",
        "name": "__Host-analytics_session",
        "type": "apiKey"
      }
    }
  },
  "info": {
    "contact": {},
    "description": "Tenant-safe operational and sales ingestion API",
    "title": "Analytics Admin API",
    "version": "0.1.0"
  },
  "openapi": "3.0.0",
  "paths": {
...
    "/api/v1/organizations/{organizationId}/governance": {
      "get": {
        "operationId": "GovernanceController_overview_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "invitations": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "acceptedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "email": { "format": "email", "type": "string" },
                          "expiresAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "id": { "format": "uuid", "type": "string" },
                          "role": {
                            "enum": ["ADMIN", "ANALYST", "VIEWER"],
                            "type": "string"
                          },
                          "status": {
                            "enum": [
                              "accepted",
                              "expired",
                              "pending",
                              "revoked"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "createdAt",
                          "email",
                          "expiresAt",
                          "id",
                          "role",
                          "status"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "jobs": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "completedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "downloadReady": { "type": "boolean" },
                          "expiresAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "failureCode": { "type": "string" },
                          "id": { "format": "uuid", "type": "string" },
                          "processedCount": { "minimum": 0, "type": "integer" },
                          "requestedByDisplayName": { "type": "string" },
                          "status": {
                            "enum": [
                              "queued",
                              "processing",
                              "completed",
                              "failed"
                            ],
                            "type": "string"
                          },
                          "subjectExternalId": {
                            "maxLength": 160,
                            "type": "string"
                          },
                          "type": {
                            "enum": [
                              "retention",
                              "privacy-export",
                              "privacy-delete"
                            ],
                            "type": "string"
                          }
                        },
                        "required": [
                          "createdAt",
                          "downloadReady",
                          "id",
                          "processedCount",
                          "requestedByDisplayName",
                          "status",
                          "type"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "settings": {
                      "additionalProperties": false,
                      "properties": {
                        "auditRetentionDays": {
                          "maximum": 3650,
                          "minimum": 30,
                          "type": "integer"
                        },
                        "exportRetentionHours": {
                          "maximum": 168,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "name": {
                          "maxLength": 100,
                          "minLength": 2,
                          "type": "string"
                        },
                        "operationalDataRetentionDays": {
                          "maximum": 3650,
                          "minimum": 30,
                          "type": "integer"
                        },
                        "privacyExportRetentionHours": {
                          "maximum": 168,
                          "minimum": 1,
                          "type": "integer"
                        },
                        "reportingTimezone": {
                          "maxLength": 80,
                          "minLength": 1,
                          "type": "string"
                        },
                        "slug": { "type": "string" },
                        "updatedAt": {
                          "format": "date-time",
                          "type": "string"
                        },
                        "weekStartsOn": {
                          "maximum": 6,
                          "minimum": 0,
                          "type": "integer"
                        }
                      },
                      "required": [
                        "auditRetentionDays",
                        "exportRetentionHours",
                        "name",
                        "operationalDataRetentionDays",
                        "privacyExportRetentionHours",
                        "reportingTimezone",
                        "slug",
                        "weekStartsOn"
                      ],
                      "type": "object"
                    },
                    "sources": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "credentialRotatedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "eventCount": { "minimum": 0, "type": "integer" },
                          "id": { "type": "string" },
                          "kind": {
                            "enum": [
                              "csv",
                              "inventory-api",
                              "orders-api",
                              "orders-webhook"
                            ],
                            "type": "string"
                          },
                          "lastReceivedAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "name": { "type": "string" },
                          "status": {
                            "enum": ["active", "revoked", "waiting"],
                            "type": "string"
                          }
                        },
                        "required": [
                          "eventCount",
                          "id",
                          "kind",
                          "name",
                          "status"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    }
                  },
                  "required": ["invitations", "jobs", "settings", "sources"],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/audit-events": {
      "get": {
        "operationId": "GovernanceController_auditEvents_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "filters": {
                      "additionalProperties": false,
                      "type": "object"
                    },
                    "items": {
                      "items": {
                        "additionalProperties": false,
                        "properties": {
                          "actorDisplayName": { "type": "string" },
                          "createdAt": {
                            "format": "date-time",
                            "type": "string"
                          },
                          "eventType": { "type": "string" },
                          "id": { "type": "string" },
                          "metadata": {
                            "additionalProperties": true,
                            "type": "object"
                          },
                          "targetId": { "type": "string" },
                          "targetType": { "type": "string" }
                        },
                        "required": [
                          "createdAt",
                          "eventType",
                          "id",
                          "metadata"
                        ],
                        "type": "object"
                      },
                      "type": "array"
                    },
                    "pageInfo": {
                      "additionalProperties": false,
                      "properties": {
                        "hasNextPage": { "type": "boolean" },
                        "nextCursor": { "type": "string" },
                        "pageSize": {
                          "maximum": 100,
                          "minimum": 1,
                          "type": "integer"
                        }
                      },
                      "required": ["hasNextPage", "pageSize"],
                      "type": "object"
                    }
                  },
                  "required": ["filters", "items", "pageInfo"],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/invitations": {
      "get": {
        "operationId": "GovernanceController_invitations_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "additionalProperties": false,
                    "properties": {
                      "acceptedAt": { "format": "date-time", "type": "string" },
                      "createdAt": { "format": "date-time", "type": "string" },
                      "email": { "format": "email", "type": "string" },
                      "expiresAt": { "format": "date-time", "type": "string" },
                      "id": { "format": "uuid", "type": "string" },
                      "role": {
                        "enum": ["ADMIN", "ANALYST", "VIEWER"],
                        "type": "string"
                      },
                      "status": {
                        "enum": ["accepted", "expired", "pending", "revoked"],
                        "type": "string"
                      }
                    },
                    "required": [
                      "createdAt",
                      "email",
                      "expiresAt",
                      "id",
                      "role",
                      "status"
                    ],
                    "type": "object"
                  },
                  "type": "array"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      },
      "post": {
        "operationId": "GovernanceController_createInvitation_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "email": { "format": "email", "type": "string" },
                  "role": {
                    "enum": ["ADMIN", "ANALYST", "VIEWER"],
                    "type": "string"
                  }
                },
                "required": ["email", "role"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "acceptancePath": { "type": "string" },
                    "acceptedAt": { "format": "date-time", "type": "string" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "email": { "format": "email", "type": "string" },
                    "expiresAt": { "format": "date-time", "type": "string" },
                    "id": { "format": "uuid", "type": "string" },
                    "role": {
                      "enum": ["ADMIN", "ANALYST", "VIEWER"],
                      "type": "string"
                    },
                    "status": {
                      "enum": ["accepted", "expired", "pending", "revoked"],
                      "type": "string"
                    }
                  },
                  "required": [
                    "createdAt",
                    "email",
                    "expiresAt",
                    "id",
                    "role",
                    "status",
                    "acceptancePath"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/invitations/{invitationId}": {
      "delete": {
        "operationId": "GovernanceController_revokeInvitation_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "invitationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "204": { "description": "" } },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/jobs": {
      "post": {
        "operationId": "GovernanceController_enqueueJob_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "subjectExternalId": {
                    "maxLength": 160,
                    "minLength": 1,
                    "type": "string"
                  },
                  "type": {
                    "enum": ["retention", "privacy-export", "privacy-delete"],
                    "type": "string"
                  }
                },
                "required": ["type"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "202": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "completedAt": { "format": "date-time", "type": "string" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "downloadReady": { "type": "boolean" },
                    "expiresAt": { "format": "date-time", "type": "string" },
                    "failureCode": { "type": "string" },
                    "id": { "format": "uuid", "type": "string" },
                    "processedCount": { "minimum": 0, "type": "integer" },
                    "requestedByDisplayName": { "type": "string" },
                    "status": {
                      "enum": ["queued", "processing", "completed", "failed"],
                      "type": "string"
                    },
                    "subjectExternalId": { "maxLength": 160, "type": "string" },
                    "type": {
                      "enum": ["retention", "privacy-export", "privacy-delete"],
                      "type": "string"
                    }
                  },
                  "required": [
                    "createdAt",
                    "downloadReady",
                    "id",
                    "processedCount",
                    "requestedByDisplayName",
                    "status",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/jobs/{jobId}": {
      "get": {
        "operationId": "GovernanceController_job_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "completedAt": { "format": "date-time", "type": "string" },
                    "createdAt": { "format": "date-time", "type": "string" },
                    "downloadReady": { "type": "boolean" },
                    "expiresAt": { "format": "date-time", "type": "string" },
                    "failureCode": { "type": "string" },
                    "id": { "format": "uuid", "type": "string" },
                    "processedCount": { "minimum": 0, "type": "integer" },
                    "requestedByDisplayName": { "type": "string" },
                    "status": {
                      "enum": ["queued", "processing", "completed", "failed"],
                      "type": "string"
                    },
                    "subjectExternalId": { "maxLength": 160, "type": "string" },
                    "type": {
                      "enum": ["retention", "privacy-export", "privacy-delete"],
                      "type": "string"
                    }
                  },
                  "required": [
                    "createdAt",
                    "downloadReady",
                    "id",
                    "processedCount",
                    "requestedByDisplayName",
                    "status",
                    "type"
                  ],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
    "/api/v1/organizations/{organizationId}/governance/jobs/{jobId}/download": {
      "get": {
        "operationId": "GovernanceController_download_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "path",
            "name": "jobId",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "responses": { "200": { "description": "" } },
        "security": [{ "session": [] }],
        "tags": ["Governance"]
      }
    },
...
    "/api/v1/organizations/{organizationId}/invitations/accept": {
      "post": {
        "operationId": "InvitationAcceptanceController_accept_v1",
        "parameters": [
          {
            "in": "path",
            "name": "organizationId",
            "required": true,
            "schema": { "type": "string" }
          },
          {
            "in": "header",
            "name": "X-CSRF-Token",
            "required": true,
            "schema": { "type": "string" }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "additionalProperties": false,
                "properties": {
                  "token": {
                    "maxLength": 200,
                    "minLength": 40,
                    "type": "string"
                  }
                },
                "required": ["token"],
                "type": "object"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": false,
                  "properties": {
                    "id": { "format": "uuid", "type": "string" },
                    "name": { "type": "string" },
                    "role": {
                      "enum": ["OWNER", "ADMIN", "ANALYST", "VIEWER"],
                      "type": "string"
                    },
                    "slug": { "type": "string" }
                  },
                  "required": ["id", "name", "role", "slug"],
                  "type": "object"
                }
              }
            },
            "description": ""
          }
        },
        "security": [{ "session": [] }],
        "tags": ["Invitations"]
      }
    },
...
  },
  "servers": [],
...
}